[Test] Move content_parts tests into existing multimodal test file
Merge into test_serving_multimodal_tokens.py to share the server
fixture and avoid launching a second Qwen3-VL instance in CI.
